Moving out Rollbar Logic outside App
